+++ src/rpc.rs
@@ -27,6 +27,10 @@ pub enum RPCSubCommand {
         about = "Get the list of unconfirmed transactions"
     )]
     UnconfirmedTXS,
+    #[structopt(name = "block", about = "Get the block at a given height")]
+    Block { height: Option<u64> },
+    #[structopt(name = "block-results", about = "Get the block at a given height")]
+    BlockResults { height: Option<u64> },
 }
 
 pub async fn rpc_cmd_parse(terra: &Terra, cmd: RPCCommand) -> Result<()> {
@@ -43,6 +47,22 @@ pub async fn rpc_cmd_parse(terra: &Terra, cmd: RPCCommand) -> Result<()> {
         }
         RPCSubCommand::UnconfirmedTXS => {
             let resp = terra.rpc(&cmd.endpoint).unconfirmed_txs().await?;
+
+            println!("{}", serde_json::to_string(&resp)?)
+        }
+        RPCSubCommand::Block { height } => {
+            let resp = match height {
+                Some(h) => terra.rpc(&cmd.endpoint).block_at_height(h).await?,
+                None => terra.rpc(&cmd.endpoint).block().await?,
+            };
+
+            println!("{}", serde_json::to_string(&resp)?)
+        }
+        RPCSubCommand::BlockResults { height } => {
+            let resp = match height {
+                Some(h) => terra.rpc(&cmd.endpoint).block_results_at_height(h).await?,
+                None => terra.rpc(&cmd.endpoint).block_results().await?,
+            };
 
             println!("{}", serde_json::to_string(&resp)?)
         }