+++ terra-rust-api/Changelog.md
@@ -1,6 +1,9 @@
 # Changes
 
 ## 1.1
+### 1.1.3 - 16-Dec-21
+* RPC Block/Block_Result api
+* New deserializer for base64 encoded strings, base64_encoded_format & base64_opt_encoded_format
 ### 1.1.2 - 14-Dec-21
 * parse IBC coins
 * tx.get_txs_in_block() function